requirements/base.txt
# Base PyTorch Testing Framework Requirements
# Core dependencies for all workloads

# PyTorch and related
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Configuration and utilities
pyyaml>=6.0
numpy>=1.24.0
pandas>=2.0.0

# Monitoring and logging
tqdm>=4.65.0
psutil>=5.9.0
py3nvml>=0.2.7

# Visualization (optional but useful)
matplotlib>=3.7.0
seaborn>=0.12.0

# Note: NVIDIA Apex is optional and must be installed from source
# https://github.com/NVIDIA/apex
# PyTorch 2.0+ has built-in AMP support (torch.cuda.amp)
